+++ srcs/builtin_echo.c
@@ -1,65 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtins_utils.c                                   :+:      :+:    :+:   */
+/*   builtin_echo.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: asyed <marvin@42.fr>                       +#+  +:+       +#+        */
+/*   By: asyed <asyed@student.42singapore.sg>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/30 14:39:14 by asyed             #+#    #+#             */
-/*   Updated: 2024/06/16 14:47:18 by asyed            ###   ########.fr       */
+/*   Updated: 2024/06/19 16:33:45 by asyed            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-int	builtin_cd(char **args, t_env **env_list)
-{
-	int		ac;
-	char	*home;
-
-	ac = 0;
-	while (args[ac])
-		ac++;
-	if (args[1] == NULL)
-	{
-		home = find_env(env_list, "HOME");
-		if (chdir(home) != 0)
-		{
-			print_error("No such file or directory");
-			return (1);
-		}
-	}
-	else if (ac > 2)
-	{
-		//printf(C "shell@st42:$ " RST);
-		print_error("Too many arguments");
-		return (1);
-	}
-	else
-	{
-		if (chdir(args[1]) != 0)
-		{
-			print_error("No such file or directory");
-			return (1);
-		}
-	}
-	return (0);
-}
-
-int	builtin_pwd(void)
-
-{
-	char cwd[4086];
-
-	if (getcwd(cwd, sizeof(cwd)) != NULL)
-		printf("%s\n", cwd);
-	else
-	{
-		perror("pwd");
-		return (1);
-	}
-	return (0);	
-}
 
 int		builtin_echo(char **args)
