+++ srcs/minishell.c
@@ -6,7 +6,7 @@
 /*   By: asyed <asyed@student.42singapore.sg>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/25 13:58:54 by ting              #+#    #+#             */
-/*   Updated: 2024/06/19 16:08:16 by asyed            ###   ########.fr       */
+/*   Updated: 2024/06/19 18:23:34 by asyed            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,40 +25,6 @@ void	signal_handlers_setup()
 	sa_int.sa_handler = sigint_handler;
 }
 */
-
-t_ms_state	*init_status()
-
-{
-	t_ms_state	*status;
-
-	status = (t_ms_state *)malloc(sizeof(t_ms_state));
-	if (status == NULL)
-	{
-		perror ("status");
-		return (NULL);
-	}
-	status->exit_status = 0;
-	return (status);
-}
-
-t_env	**init_status_and_envdup(t_ms_state **status, char **env)
-
-{
-	t_env	**env_dup;
-
-	*status = init_status();
-	if (status == NULL)
-		return (NULL);
-	env_dup = init_env_copy(env);
-	if (env_dup == NULL)
-	{
-		perror("init_env_copy");
-		free(*status);
-		*status = NULL;
-		return(NULL);
-	}
-	return (env_dup);
-}
 
 int	main(int ac, char **av, char **env)
 {
@@ -100,7 +66,7 @@ int	main(int ac, char **av, char **env)
 			while(current)
 			{
 				if (current->builtin)
-					status->exit_status = execute_builtins(cmds, current->cmd_arr, env_dup, status);
+					execute_builtins(cmds, current->cmd_arr, env_dup, status);
 				current = current->next;
 			}
 			free_cmds(cmds);
